requirements.txt
# local package
-e .

# external requirements
click~=7.1.2
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

requests~=2.25.1
setuptools~=56.0.0
python-dateutil~=2.8.1